ip, doc, authors
- `core.object.ip6`: Fix man-page newline
- `lib.ip`: New library, move IP address printing functions from core objects
- `lib.trie`: Fix spacing in synopsis code
- Add Tomas Krizek to authors
