+++ src/lib.lua
@@ -22,6 +22,7 @@ module(...,package.seeall)
 
 -- dnsjit.lib.clock (3),
 -- dnsjit.lib.getopt (3),
+-- dnsjit.lib.ip (3),
 -- dnsjit.lib.parseconf (3),
 -- dnsjit.lib.trie (3)
 return
